+++ src/components/Popup.js
@@ -1,0 +1,45 @@
+import React, { Component, Fragment } from 'react'
+import { X } from 'react-feather'
+
+import './Popup.css'
+
+class Popup extends Component {
+  constructor(props) {
+    super(props)
+    this.state = { showPopup: false }
+  }
+
+  togglePopup() {
+    this.setState({
+      showPopup: !this.state.showPopup
+    })
+  }
+
+  render() {
+    const { children } = this.props
+    return (
+      <Fragment>
+        <div className="taCenter">
+          <h3> Simple Popup Example</h3>
+          <div class="Button" onClick={this.togglePopup.bind(this)}>
+            Click To Launch Popup
+          </div>
+        </div>
+
+        {this.state.showPopup ? (
+          <div className="Popup-Overlay">
+            <div
+              className="Popup-Background"
+              onClick={this.togglePopup.bind(this)}
+            ></div>
+            <div className="Popup-Inner">
+              <X class="Popup-Close" onClick={this.togglePopup.bind(this)} />
+              {children}
+            </div>
+          </div>
+        ) : null}
+      </Fragment>
+    )
+  }
+}
+export default Popup