+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+## 2.0.10 - 2019-08-01
+
+Change Accordion click event to toggle individual item.
+
+## 2.0.9 - 2019-07-26
+
+Add new Popup component - Accepts children :)
+
 ## 2.0.9 - 2019-05-17
 
 Add google tag manager to track routes